Nugetize TensorFlowSharp
Adds the package metadata and native files to the project.

The native files are packaged explicitly because contentFiles with
BundleResource build action aren't working for me (nuget restore
issue, it seems). So an explicit MSBuild targets is also packaged
that references the files in the package location.

Also removed the packages.config approach and switched to
PackageReference so that you can always build against the latest
and greatest nugetizer as it becomes available.